packages/ai/src/ai/common/models/audio/requirements_piper.txt
huggingface_hub>=0.20.0
piper-tts>=1.2.0,<2
lameenc>=1.6.0
imageio-ffmpeg>=0.4.0
